refactor(view): add custom Hook ClusterGraph draw logic split
